Move Add to Calendar button to CalendarNav
