+++ packages/client/src/pages/customer_area/index.tsx
@@ -14,9 +14,14 @@ import {
 import BookView from "./views/Book";
 import CalendarView from "./views/Calendar";
 import { NotificationsContainer } from "@/features/notifications/components";
+import AddToCalendar from "@/components/atoms/AddToCalendar";
 
 import useFirestoreSubscribe from "@/react-redux-firebase/hooks/useFirestoreSubscribe";
-import { getBookingsCustomer } from "@/store/selectors/bookings";
+import {
+  getBookedSlotsByMonth,
+  getBookingsCustomer,
+  getSlotsForCustomer,
+} from "@/store/selectors/bookings";
 import { getCalendarDay } from "@/store/selectors/app";
 import { getIsAdmin } from "@/store/selectors/auth";
 import { changeCalendarDate } from "@/store/actions/appActions";
@@ -32,6 +37,10 @@ const CustomerArea: React.FC = () => {
   useSecretKey();
 
   const isAdmin = useSelector(getIsAdmin);
+  const date = useSelector(getCalendarDay);
+  const slotsByDay = useSelector(getSlotsForCustomer);
+
+  const bookedSlotsByMonth = useSelector(getBookedSlotsByMonth(date.month));
 
   // Subscribe to necessary collections
   useFirestoreSubscribe([
@@ -77,6 +86,12 @@ const CustomerArea: React.FC = () => {
     </>
   );
 
+  const addToCalendar = Object.keys(bookedSlotsByMonth).length ? (
+    <AddToCalendar bookedSlots={bookedSlotsByMonth} slots={slotsByDay} />
+  ) : (
+    <></>
+  );
+
   const CustomerView = views[view];
 
   return (
@@ -87,7 +102,11 @@ const CustomerArea: React.FC = () => {
       additionalButtons={additionalButtons}
       user={displayCustomer}
     >
-      <CalendarNav {...calendarNavProps} jump="month" />
+      <CalendarNav
+        {...calendarNavProps}
+        additionalContent={addToCalendar}
+        jump="month"
+      />
       <div className="content-container">
         <div className="px-[44px] py-4">
           <CustomerView />
